pyproject.toml
[project]
name = "electronics-trade-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.5",
    "django-filter>=24.3",
    "django-stubs>=5.1.2",
    "djangorestframework>=3.15.2",
    "flake8-pyproject>=1.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]

[tool.flake8]
max-line-length = 120
exclude = ['.venv', 'migrations']
